change validFrom/validUntil column data type to int
Summary:
change validFrom/until data type from long to int, change memory usage per document from
16 bytes to 8 bytes

Reviewers: bzzhang, csoman, tingchen

Reviewed By: bzzhang

Differential Revision: https://code.uberinternal.com/D3534725

+++ pinot-core/src/main/java/org/apache/pinot/core/indexsegment/immutable/ImmutableUpsertSegmentImpl.java
@@ -150,7 +150,7 @@ public class ImmutableUpsertSegmentImpl extends ImmutableSegmentImpl implements 
     int docId = getDocIdFromSourceOffset(offset);
     StringBuilder result = new StringBuilder("matched: ");
     for (VirtualColumnLongValueReaderWriter readerWriter : _virtualColumnsReaderWriter) {
-      result.append(readerWriter.getLong(docId)).append("; ");
+      result.append(readerWriter.getInt(docId)).append("; ");
     }
     return result.toString();
   }
